logic 1.5
