+++ logic/menu.py
@@ -20,11 +20,16 @@ def load_menu(screen, clock):
             start_btn_color = (0, 0, 0)
             if event.type == pygame.QUIT:
                 terminate()
+            elif event.type == pygame.KEYDOWN or \
+                    event.type == pygame.MOUSEBUTTONDOWN:
+                return
             try:
                 if start_btn.rect.collidepoint(event.pos):
                     start_btn_color = (255, 0, 0)
+
             except AttributeError:
                 pass
+
         start_btn = Buttons('Start', start_btn_color, (logic.constants.WIDTH // 2, logic.constants.HEIGHT // 2))
 
         screen.blit(start_btn.text,
