+++ logic/player.py
@@ -1,6 +1,6 @@
 import pygame
 
-from load_image import *
+from logic.load_image import *
 
 player_group = pygame.sprite.Group()
